Per-tool search
